+++ src/allreduce/allreduce.h
@@ -53,6 +53,6 @@ extern StopWatch all_reduce_watch;
 extern long long all_reduce_data_amount;
 extern int all_reduce_counter;
 
-void all_reduce(double* buffer, int n, std::string master_location, size_t unique_id, size_t total, size_t node, node_socks& socks);
+void all_reduce(float* buffer, int n, std::string master_location, size_t unique_id, size_t total, size_t node, node_socks& socks);
 }
 #endif
